requirements.txt
absl-py==2.0.0
astunparse==1.6.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.47.0
gast==0.5.4
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h5py==3.10.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.6
imageio==2.33.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==9.5.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyMatting==1.1.12
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
referencing==0.32.0
rembg==2.0.53
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.15.2
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
tifffile==2023.12.9
tqdm==4.66.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
